Dev/oisnull/group

+++ app/src/main/java/com/hwl/beta/sp/UserPosSP.java
@@ -16,29 +16,29 @@ public class UserPosSP {
     private static final String USERPOSID = "userposid";
     private static final String GROUPGUID = "groupguid";
     private static final String LATITUDE = "latitude";
-    private static final String LONTITUDE = "lontitude";
+    private static final String LONGITUDE = "longitude";
     private static final String COUNTRY = "country";
     private static final String PROVINCE = "province";
     private static final String CITY = "city";
     private static final String DISTRICT = "district";
     private static final String STREET = "street";
     private static final String ADDR = "addr";
-    private static final String LASTNEARCIRCLEID = "lastNearCircleId";
+//    private static final String LASTNEARCIRCLEID = "lastNearCircleId";
 
     private static SharedPreferences getSP() {
         return HWLApp.getContext().getSharedPreferences(USERPOSPREFERENCE, Context.MODE_PRIVATE);
     }
 
-    public static boolean isExistsPosInfo() {
-        if (getLatitude() < 0 || getLontitude() < 0) {
+    public static boolean isExistPosInfo() {
+        if (getLatitude() < 0 || getLongitude() < 0) {
             return false;
         }
         return true;
     }
 
-    public static void setUserPos(float latitude, float lontitude, String country, String
+    public static void setUserPos(float latitude, float longitude, String country, String
             province, String city, String district, String street, String addr) {
-        setUserPos(0, "", latitude, lontitude, country, province, city, district, street, addr);
+        setUserPos(0, "", latitude, longitude, country, province, city, district, street, addr);
     }
 
     public static void setUserPos(int userPosId, String groupGuid) {
@@ -48,15 +48,22 @@ public class UserPosSP {
         editor.commit();
     }
 
-    public static void setUserPos(int userPosId, String groupGuid, float latitude, float
-            lontitude, String country, String province, String city, String district, String
-            street, String addr) {
+    public static void setUserPos(int userPosId,
+                                  String groupGuid,
+                                  double latitude,
+                                  double longitude,
+                                  String country,
+                                  String province,
+                                  String city,
+                                  String district,
+                                  String street,
+                                  String addr) {
         final SharedPreferences.Editor editor = getSP().edit();
         editor.putInt(USERPOSID, userPosId);
         editor.putString(GROUPGUID, groupGuid);
 
-        editor.putFloat(LATITUDE, latitude);
-        editor.putFloat(LONTITUDE, lontitude);
+        editor.putFloat(LATITUDE, (float) latitude);
+        editor.putFloat(LONGITUDE, (float) longitude);
 
         editor.putString(COUNTRY, country);
         editor.putString(PROVINCE, province);
@@ -67,18 +74,18 @@ public class UserPosSP {
         editor.commit();
     }
 
-    public static void setLastNearCircleId(long lastNearCircleId) {
-        if (getLastNearCircleId() >= lastNearCircleId) {
-            return;
-        }
-        final SharedPreferences.Editor editor = getSP().edit();
-        editor.putLong(LASTNEARCIRCLEID, lastNearCircleId);
-        editor.commit();
-    }
+//    public static void setLastNearCircleId(long lastNearCircleId) {
+//        if (getLastNearCircleId() >= lastNearCircleId) {
+//            return;
+//        }
+//        final SharedPreferences.Editor editor = getSP().edit();
+//        editor.putLong(LASTNEARCIRCLEID, lastNearCircleId);
+//        editor.commit();
+//    }
 
-    public static long getLastNearCircleId() {
-        return getSP().getLong(LASTNEARCIRCLEID, 0);
-    }
+//    public static long getLastNearCircleId() {
+//        return getSP().getLong(LASTNEARCIRCLEID, 0);
+//    }
 
     public static int getUserPosId() {
         return getSP().getInt(USERPOSID, 0);
@@ -92,8 +99,8 @@ public class UserPosSP {
         return getSP().getFloat(LATITUDE, -1);
     }
 
-    public static float getLontitude() {
-        return getSP().getFloat(LONTITUDE, -1);
+    public static float getLongitude() {
+        return getSP().getFloat(LONGITUDE, -1);
     }
 
     public static String getPosDesc() {
@@ -126,8 +133,26 @@ public class UserPosSP {
     }
 
     public static void clearPosInfo() {
-        final SharedPreferences.Editor editor = getSP().edit();
-        editor.clear();
+        clearPosInfo(false);
+    }
+
+    public static void clearPosInfo(boolean keepGroup) {
+        SharedPreferences.Editor editor = getSP().edit();
+        if (keepGroup) {
+            editor.putInt(USERPOSID, 0);
+//            editor.putString(GROUPGUID, groupGuid);
+            editor.putFloat(LATITUDE, 0);
+            editor.putFloat(LONGITUDE, 0);
+
+            editor.putString(COUNTRY, null);
+            editor.putString(PROVINCE, null);
+            editor.putString(CITY, null);
+            editor.putString(DISTRICT, null);
+            editor.putString(STREET, null);
+            editor.putString(ADDR, null);
+        } else {
+            editor.clear();
+        }
         editor.commit();
     }
 }
